+++ test/fixtures/stress/recovery/void_closers.blade.php
@@ -1,0 +1,8 @@
+@if($show)
+<section>
+  <br></br>
+  <input type="text"></input>
+  <img></img>
+  <p>Missing closers</p>
+</section>
+</section>